Feature tbcapi v00003

+++ entity/electrumx/electrumx.go
@@ -3,7 +3,7 @@ package electrumx
 // ElectrumXHistoryItem 表示单个交易历史记录项
 type ElectrumXHistoryItem struct {
 	TxHash string `json:"tx_hash"`
-	Height int    `json:"height"`
+	Height int64    `json:"height"`
 }
 
 // ElectrumXHistoryResponse 表示从ElectrumX获取的历史记录响应
